// apps/desktop/src/ipc/generated/RequestState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Where a user's request stands.
 */
export type RequestState = { "type": "working" } | { "type": "waiting" } | { "type": "done" } | { "type": "stopped" } | { "type": "failed", error: string, };